da-tools-33-add-an-exemple-with-quasi-geostrophic-equations/requirements.txt
tensorboard
rich
coverage
pylint
setuptools_scm
torch-tb-profiler
pytorch-lightning
tensordict
pytest
pytest-cov
torchmetrics
torchvision
gitpython
black
h5py
git+https://codebase.helmholtz.cloud/m-dml/mdml-sim@main

sphinx<7.0.0
sphinx-autoapi
sphinx-typo3-theme # sphinx theme
seed-intersphinx-mapping
